Removed main_hot.h and renamed HOT to SingleNodeMortonHOT.
