+++ cpp/include/tapas.h
@@ -4,7 +4,6 @@
 // Base header for user code. 
 
 #include "tapas/main.h"
-#include "tapas/main_hot.h"
 #include "tapas/map.h"
 
 #endif
